Add lint method check
