+++ mapbox-lint/src/main/java/com/mapbox/maps/lint/LifecycleIssueRegistry.kt
@@ -5,7 +5,7 @@ import com.android.tools.lint.client.api.IssueRegistry
 class LifecycleIssueRegistry : IssueRegistry() {
 
   override val issues = listOf(
-      MapboxObsoleteDetector.ISSUE,MethodDetector.ISSUE)
+      LifecycleMethodDetector.ISSUE)
 
   override val api: Int = com.android.tools.lint.detector.api.CURRENT_API
 }
