--- mapbox-lint/src/test/java/com/mapbox/maps/lint/MethodDetectorTest.kt
@@ -1,213 +1,0 @@
-package com.mapbox.maps.lint
-
-import com.android.tools.lint.checks.infrastructure.TestFiles
-import com.android.tools.lint.checks.infrastructure.TestLintTask
-import org.junit.Test
-
-class MethodDetectorTest {
-
-  @Test
-  fun mapViewOnStart() {
-    TestLintTask.lint().files(
-      TestFiles.java(
-        """
-                    package test.pkg;
-                    import com.mapbox.maps.MapView;
-                    public class TestClass {
-                        MapView mapView = new MapView();
-                        void onStart() {
-                            mapView.onStart();
-                        }
-                    }
-                    """
-      ).indented()
-    )
-
-      .issues(
-        MethodDetector.ISSUE
-      )
-      .run()
-      .expect(
-        """
-         |src/test/pkg/TestClass.java:6: Warning: No need to invoke onStart/onStop/onDestroy/onLowMemory with Mapbox Lifecycle Plugin. [Lifecycle]
-         |        mapView.onStart();
-         |        ~~~~~~~~~~~~~~~~~
-         |0 errors, 1 warnings
-        """.trimMargin()
-      )
-  }
-
-  @Test
-  fun mapViewOnStop() {
-    TestLintTask.lint().files(
-      TestFiles.java(
-        """
-                    package test.pkg;
-                    import com.mapbox.maps.MapView;
-                    public class TestClass {
-                        MapView mapView = new MapView();
-                        void onStop() {
-                            mapView.onStop();
-                        }
-                    }
-                    """
-      ).indented()
-    )
-      .issues(
-        MethodDetector.ISSUE
-      )
-      .run()
-      .expect(
-        """
-         |src/test/pkg/TestClass.java:6: Warning: No need to invoke onStart/onStop/onDestroy/onLowMemory with Mapbox Lifecycle Plugin. [Lifecycle]
-         |        mapView.onStop();
-         |        ~~~~~~~~~~~~~~~~
-         |0 errors, 1 warnings
-        """.trimMargin()
-      )
-  }
-
-  @Test
-  fun mapViewOnDestroy() {
-    TestLintTask.lint().files(
-      TestFiles.java(
-        """
-                    package test.pkg;
-                    import com.mapbox.maps.MapView;
-                    public class TestClass {
-                        MapView mapView = new MapView();
-                        void onDestroy() {
-                            mapView.onDestroy();
-                        }
-                    }
-                    """
-      ).indented()
-    )
-      .issues(
-        MethodDetector.ISSUE
-      )
-      .run()
-      .expect(
-        """
-         |src/test/pkg/TestClass.java:6: Warning: No need to invoke onStart/onStop/onDestroy/onLowMemory with Mapbox Lifecycle Plugin. [Lifecycle]
-         |        mapView.onDestroy();
-         |        ~~~~~~~~~~~~~~~~~~~
-         |0 errors, 1 warnings
-        """.trimMargin()
-      )
-  }
-
-  @Test
-  fun mapViewOnLowMemory() {
-    TestLintTask.lint().files(
-      TestFiles.java(
-        """
-                    package test.pkg;
-                    import com.mapbox.maps.MapView;
-                    public class TestClass {
-                        MapView mapView = new MapView();
-                        void onLowMemory() {
-                            mapView.onLowMemory();
-                        }
-                    }
-                    """
-      ).indented()
-    )
-      .issues(
-        MethodDetector.ISSUE
-      )
-      .run()
-      .expect(
-        """
-         |src/test/pkg/TestClass.java:6: Warning: No need to invoke onStart/onStop/onDestroy/onLowMemory with Mapbox Lifecycle Plugin. [Lifecycle]
-         |        mapView.onLowMemory();
-         |        ~~~~~~~~~~~~~~~~~~~~~
-         |0 errors, 1 warnings
-        """.trimMargin()
-      )
-  }
-
-  @Test
-  fun normalOnStart() {
-    TestLintTask.lint().files(
-      TestFiles.java(
-        """
-                    package test.pkg;
-                    public class TestClass {
-                        void onStart() {
-                            foo.onStart();
-                        }
-                    }
-                    """
-      ).indented()
-    )
-      .issues(
-        MethodDetector.ISSUE
-      )
-      .run()
-      .expectClean()
-  }
-
-  @Test
-  fun normalOnStop() {
-    TestLintTask.lint().files(
-      TestFiles.java(
-        """
-                    package test.pkg;
-                    public class TestClass {
-                        void onStop() {
-                            foo.onStop();
-                        }
-                    }
-                    """
-      ).indented()
-    )
-      .issues(
-        MethodDetector.ISSUE
-      )
-      .run()
-      .expectClean()
-  }
-
-  @Test
-  fun normalOnDestroy() {
-    TestLintTask.lint().files(
-      TestFiles.java(
-        """
-                    package test.pkg;
-                    public class TestClass {
-                        void onDestroy() {
-                            foo.onDestroy();
-                        }
-                    }
-                    """
-      ).indented()
-    )
-      .issues(
-        MethodDetector.ISSUE
-      )
-      .run()
-      .expectClean()
-  }
-
-  @Test
-  fun normalOnLowMemory() {
-    TestLintTask.lint().files(
-      TestFiles.java(
-        """
-                    package test.pkg;
-                    public class TestClass {
-                        void onLowMemory() {
-                            foo.onLowMemory();
-                        }
-                    }
-                    """
-      ).indented()
-    )
-      .issues(
-        MethodDetector.ISSUE
-      )
-      .run()
-      .expectClean()
-  }
-}
