--- sdk-base/src/main/java/com/mapbox/maps/MapboxObsolete.kt
@@ -1,8 +1,0 @@
-package com.mapbox.maps
-
-/**
- * Annotation class to mark API as obsolete.
- */
-@Retention(AnnotationRetention.BINARY)
-@Target(AnnotationTarget.FUNCTION)
-annotation class MapboxObsolete
